supporting-blog-content/building-multimodal-rag-with-elasticsearch-gotham/requirements.txt
elasticsearch>=8.11.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
imagebind @ git+https://github.com/facebookresearch/ImageBind.git
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
librosa>=0.10.0
matplotlib>=3.7.0